Refactor, change stash to hold keys
